requirements/dev.txt
-r ./production.txt
# General development
django-debug-toolbar==1.8
django-extensions
django-nose
flake8
typing
# Linting
git-lint
pylint
pep8
yamllint
html-linter
# Testing
factory_boy
coverage
django_coverage_plugin
hypothesis==3.56.5
